rename blobjects to objbatch and S3 to ObjectStorage
Improves naming to better reflect the generic object storage batching functionality.
Package objbatch handles batching for any BlobLike storage backend.
