+++ cmd/reveald/config.go
@@ -14,8 +14,8 @@ import (
 	"github.com/runreveal/kawa/x/s3"
 	"github.com/runreveal/lib/loader"
 	"github.com/runreveal/reveald/internal"
-	"github.com/runreveal/reveald/internal/destinations/blobjects"
 	mqttDstkawad "github.com/runreveal/reveald/internal/destinations/mqtt"
+	"github.com/runreveal/reveald/internal/destinations/objbatch"
 	"github.com/runreveal/reveald/internal/destinations/printer"
 	"github.com/runreveal/reveald/internal/destinations/runreveal"
 	s3kawad "github.com/runreveal/reveald/internal/destinations/s3"
@@ -69,7 +69,7 @@ func init() {
 		return &S3Config{}
 	})
 	loader.Register("s3b", func() loader.Builder[kawa.Destination[types.Event]] {
-		return &blobjects.BlobConfig{}
+		return &objbatch.BlobConfig{}
 	})
 	loader.Register("runreveal", func() loader.Builder[kawa.Destination[types.Event]] {
 		return &RunRevealConfig{}
